put deprecated scripts into separater folder
I'm now using objects for my subjects and modeling, so these scripts are old. The model is also updated, and is now the most accurate. Before, the model didn't differentiate between decision time and leave time. Decision time and leave time are different when reacting, but the same when gambling. Also, people are forced to react if their decision time is after the the agent... indecisions only come from reacting now, which makes sense, but participants are making gamble indecisions in the task... something to think about