exp2 pipeline basically completed, see xtra message for deets
The subject object, I removed all the mean and sd calculations from the object, gonna do that directly in the jupyter notebook. This makes the object cleaner and it's not that much extra typing in the notebook. Plus, I can now mask things on the fly, rather than them being stuck in the object